feat: post , delete

+++ Todo_Backend/TodoListApi/Controller/TodoController.cs
@@ -74,7 +74,7 @@ namespace TodoListApi.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteTodoItem(int id)
+        public async Task<IActionResult> DeleteTodoItem(long id)
         {
             var todoItem = await _context.TodoItems.FindAsync(id);
             if (todoItem == null)
